pyproject.toml
[project]
name = "etl-pipeline-automation-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.4.4",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langgraph[checkpoint]>=1.0.10",
    "openai>=2.25.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlglot>=29.0.1",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]
